Build Test !KeepVersion

+++ Build/build.ps1
@@ -24,15 +24,18 @@ ForEach ($ManualModuleKey in $ManualModuleKeys) {
     $ModuleName = $ManualModuleKey
     $ModuleVersion = $ManualModuleInstalls[$ManualModuleKey]
 
-    $File         = 'https://www.powershellgallery.com/api/v2/package/{0}/{1}' -f $ModuleName, $ModuleVersion
-    $ModulePath   = '\{0}\{1}\' -f $ModuleName, $ModuleVersion
-    $TempFilePath = '{0}\PSGalleryFix{1}{2}.{3}.zip' -f $env:TEMP, $ModulePath, $ModuleName, $ModuleVersion
-    $ExpandPath   = '{0}\WindowsPowerShell\Modules{1}' -f $env:ProgramFiles, $ModulePath
+    # If the current version is less than the specified version update it
+    If (((Get-Module $ModuleName -ListAvailable)[0].Version) -lt $ModuleVersion) {
+        $File         = 'https://www.powershellgallery.com/api/v2/package/{0}/{1}' -f $ModuleName, $ModuleVersion
+        $ModulePath   = '\{0}\{1}\' -f $ModuleName, $ModuleVersion
+        $TempFilePath = '{0}\PSGalleryFix{1}{2}.{3}.zip' -f $env:TEMP, $ModulePath, $ModuleName, $ModuleVersion
+        $ExpandPath   = '{0}\WindowsPowerShell\Modules{1}' -f $env:ProgramFiles, $ModulePath
 
-    $null = New-Item -Path $env:TEMP\PSGalleryFix\$ModulePath -ItemType Directory -Force
-    Invoke-WebRequest -Uri $File -UseBasicParsing -OutFile $TempFilePath
-    Expand-Archive -Path $TempFilePath -DestinationPath $ExpandPath -Force
-    Remove-Item -Path $TempFilePath
+        $null = New-Item -Path $env:TEMP\PSGalleryFix\$ModulePath -ItemType Directory -Force
+        Invoke-WebRequest -Uri $File -UseBasicParsing -OutFile $TempFilePath
+        Expand-Archive -Path $TempFilePath -DestinationPath $ExpandPath -Force
+        Remove-Item -Path $TempFilePath
+    }
 }
 
 Remove-Module PowerShellGet,PackageManagement -Force
@@ -40,7 +43,7 @@ Import-Module -Name PowerShellGet -Force
 Import-PackageProvider -Name PowerShellGet -Force -RequiredVersion $ManualModuleInstalls['PowerShellGet']
 
 # Install build dependency modules
-$Modules = @("Psake","PSDeploy","BuildHelpers","PSScriptAnalyzer","Pester","Posh-Git")
+$Modules = @("Psake","PSDeploy","BuildHelpers","Pester","Posh-Git","PSScriptAnalyzer")
 
 ForEach ($Module in $Modules) {
     If (-not (Get-Module -Name $Module -ListAvailable)) {
@@ -51,9 +54,20 @@ ForEach ($Module in $Modules) {
     }
 
     Try {
-        Import-Module $Module -ErrorAction Stop
+        Import-Module $Module -Force -ErrorAction Stop
     }
     Catch {
+        If ($Module -eq 'PSScriptAnalyzer') {
+            Write-Warning 'Attempting PSScriptAnalyzer install a second time'
+            Try {
+                Install-Module $Module -Force -ErrorAction Stop -SkipPublisherCheck
+                Start-Sleep -Seconds 5
+                Import-Module $Module -Force -ErrorAction Stop
+            }
+            Catch {
+                $PSCmdlet.ThrowTerminatingError($PSItem)
+            }
+        }
         Write-Error $PSItem
         Exit
     }
